backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-axes==6.1.1
django-cors-headers==4.3.1
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.95.0
gunicorn==21.2.0
iniconfig==2.1.0
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
dj-database-url==2.1.0
